renaming application to ApplicationAdministration
