+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/ApplicationConfiguration.tsx
