+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/ApplicationQuestionConfig/components/ApplicationQuestionCard.tsx
