+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/ApplicationQuestionConfig/components/DeleteConfirmation.tsx
