+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/ApplicationQuestionConfig/components/MultiSelectConfig.tsx
