+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/ApplicationQuestionConfig/components/TextConfig.tsx
