+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/ApplicationQuestionConfig/handlers/handleSubmitAllQuestions.ts
