+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/components/ApplicationConfigDialog.tsx
