+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/components/ConfigurationHeader.tsx
