+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/components/ExternalStudentsAllowed.tsx
