+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/interfaces/ApplicationMetaData.ts
