+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/interfaces/ApplicationStatus.ts
