+++ clients/core/src/ApplicationAdministration/ApplicationConfiguration/utils/getApplicationStatus.ts
